[CHORE] #65 : path 수정, remote/api -> service
