+++ client/app/src/main/java/com/example/healthc/data/service/KorProductService.kt
@@ -1,7 +1,7 @@
-package com.example.healthc.data.remote.api
+package com.example.healthc.data.service
 
 import com.example.healthc.BuildConfig
-import com.example.healthc.data.dto.kor_product.KorProductResponse
+import com.example.healthc.data.model.remote.kor_product.KorProductResultResponse
 import retrofit2.http.GET
 import retrofit2.http.Query
 
@@ -12,6 +12,6 @@ interface KorProductService {
         @Query("ServiceKey") ServiceKey : String = BuildConfig.DATA_GO_PRODUCT_API_KEY,
         @Query("returnType") returnType : String = "json",
         @Query("prdlstNm") prdlstNm : String
-    ) : KorProductResponse
+    ) : KorProductResultResponse
 
 }
