+++ client/app/src/main/java/com/example/healthc/data/service/ObjectDetectionService.kt
@@ -1,7 +1,7 @@
-package com.example.healthc.data.remote.api
+package com.example.healthc.data.service
 
 import com.example.healthc.BuildConfig
-import com.example.healthc.data.dto.object_detection.DetectedObjectDto
+import com.example.healthc.data.model.remote.detection.ObjectDetectionResponse
 import okhttp3.MultipartBody
 import retrofit2.http.Multipart
 import retrofit2.http.POST
@@ -13,8 +13,8 @@ interface ObjectDetectionService {
     // 음식 사진 인식
     @Multipart
     @POST("/food/images/classify")
-    suspend fun postFoodImage(
+    suspend fun getDetectedObject(
         @Query("apiKey") apiKey : String = BuildConfig.SPOON_API_KEY,
         @Part file: MultipartBody.Part,
-    ) : DetectedObjectDto
+    ) : ObjectDetectionResponse
 }
